requirements/requirements.txt
albumentations>=0.5.2,<=1.0.0
facenet_pytorch==2.5.2
hydra-core==1.1.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python>=4.5.1.48,<=4.5.5.64
pandas==1.1.5
Pillow>=8.1.2,<=8.4.0
pytorch-lightning==1.5.10
scikit-learn==0.24.1
timm>=0.3.2,<=0.5.4
torch>=1.8.0
tqdm>=4.59.0,<=4.63.0
